refactor: dissolve core/schemas.ts — all schemas now in @agents/sdk
Migrate all 13 CLI consumers from @agents/core/schemas to SDK locations:
- Lockfile/external schemas -> @agents/sdk/providers/local/schemas
- Plugin/marketplace schemas -> @agents/sdk/context/plugin/schema
- Skill frontmatter schema -> @agents/sdk/context/skill/schema

Add backward-compatible aliases (un-suffixed names) so existing consumer
code works without renaming types. Add nested subpath exports to SDK
package.json for deep imports.

Test results: core 10/10, SDK 132/132, CLI 1929/1929 (11 pre-existing failures unchanged).
