+++ packages/cli/src/lib/manifest.ts
@@ -8,15 +8,15 @@
 import { existsSync } from 'node:fs'
 import { readFile } from 'node:fs/promises'
 import { join } from 'node:path'
+import { CliError, err, ok, type Result, type SourceFormat } from '@agents/core/types'
 import {
   PluginManifest,
   type PluginSource,
   PluginSourceExtended,
   PluginSourcePlanning,
   PluginSourcesManifest,
-  SkillFrontmatter,
-} from '@agents/core/schemas'
-import { CliError, err, ok, type Result, type SourceFormat } from '@agents/core/types'
+} from '@agents/sdk/context/plugin/schema'
+import { SkillFrontmatter } from '@agents/sdk/context/skill/schema'
 import * as v from 'valibot'
 import { parseFrontmatter } from './chunker'
