+++ packages/core/src/index.ts
@@ -2,7 +2,7 @@
  * @agents/core — Shared foundation for AI Context Library tools.
  *
  * Provides: Result monad, file I/O, config, output formatting,
- * component type system, git/GitHub integration, schemas, hashing.
+ * component type system, git/GitHub integration, hashing.
  *
  * This barrel re-exports all public APIs. For tree-shaking or
  * smaller imports, use deep paths: `@agents/core/types`, etc.
@@ -16,7 +16,6 @@ export * from './git'
 export * from './github'
 export * from './hash'
 export * from './runtime'
-export * from './schemas'
 export * from './source-parser'
 // Foundation
 export * from './types'
