+++ packages/sdk/package.json
@@ -10,8 +10,20 @@
     "./util/*": "./src/util/*.ts",
     "./context": "./src/context/index.ts",
     "./context/*": "./src/context/*.ts",
+    "./context/plugin/*": "./src/context/plugin/*.ts",
+    "./context/skill/*": "./src/context/skill/*.ts",
+    "./context/agent/*": "./src/context/agent/*.ts",
+    "./context/command/*": "./src/context/command/*.ts",
+    "./context/hook/*": "./src/context/hook/*.ts",
+    "./context/mcp/*": "./src/context/mcp/*.ts",
+    "./context/rule/*": "./src/context/rule/*.ts",
+    "./context/output-style/*": "./src/context/output-style/*.ts",
     "./providers": "./src/providers/index.ts",
     "./providers/*": "./src/providers/*.ts",
+    "./providers/local/*": "./src/providers/local/*.ts",
+    "./providers/github/*": "./src/providers/github/*.ts",
+    "./providers/smithery/*": "./src/providers/smithery/*.ts",
+    "./providers/clients/*": "./src/providers/clients/*.ts",
     "./catalog": "./src/catalog/index.ts",
     "./catalog/*": "./src/catalog/*.ts",
     "./ui": "./src/ui/index.ts",
